+++ components/blockeditor/src/blockeditor/index.ts
@@ -1,2 +1,1 @@
-export { BlocksDirective, BlockDirective, BlocksPlugin, BlockPlugin } from './blocks.directive';
 export { BlockEditorComponent, BlockEditorPlugin } from './blockeditor.component';