+++ components/kanban/src/kanban/kanban.component.ts
@@ -6,7 +6,7 @@ import { ColumnsDirective, ColumnDirective, ColumnsPlugin, ColumnPlugin } from '
 import { StackedHeadersDirective, StackedHeaderDirective, StackedHeadersPlugin, StackedHeaderPlugin } from './stackedheaders.directive'
 
 
-export const properties: string[] = ['isLazyUpdate', 'plugins', 'allowDragAndDrop', 'allowKeyboard', 'cardHeight', 'cardSettings', 'columns', 'constraintType', 'cssClass', 'dataSource', 'dialogSettings', 'enableHtmlSanitizer', 'enablePersistence', 'enableRtl', 'enableTooltip', 'enableVirtualization', 'externalDropId', 'height', 'keyField', 'locale', 'query', 'showEmptyColumn', 'sortSettings', 'stackedHeaders', 'swimlaneSettings', 'tooltipTemplate', 'width', 'actionBegin', 'actionComplete', 'actionFailure', 'cardClick', 'cardDoubleClick', 'cardRendered', 'created', 'dataBinding', 'dataBound', 'dataSourceChanged', 'dataStateChange', 'dialogClose', 'dialogOpen', 'drag', 'dragStart', 'dragStop', 'queryCellInfo'];
+export const properties: string[] = ['isLazyUpdate', 'plugins', 'allowColumnDragAndDrop', 'allowDragAndDrop', 'allowKeyboard', 'cardHeight', 'cardSettings', 'columns', 'constraintType', 'cssClass', 'dataSource', 'dialogSettings', 'enableHtmlSanitizer', 'enablePersistence', 'enableRtl', 'enableTooltip', 'enableVirtualization', 'externalDropId', 'height', 'keyField', 'locale', 'query', 'showEmptyColumn', 'sortSettings', 'stackedHeaders', 'swimlaneSettings', 'tooltipTemplate', 'width', 'actionBegin', 'actionComplete', 'actionFailure', 'cardClick', 'cardDoubleClick', 'cardRendered', 'columnDrag', 'columnDragStart', 'columnDrop', 'created', 'dataBinding', 'dataBound', 'dataSourceChanged', 'dataStateChange', 'dialogClose', 'dialogOpen', 'drag', 'dragStart', 'dragStop', 'queryCellInfo'];
 export const modelProps: string[] = [];
 
 export const testProp: any = getProps({props: properties});
@@ -154,8 +154,14 @@ export let KanbanComponent: DefineVueComponent<KanbanModel> =  vueDefineComponen
         templateParser(template: string | Object): any {
             return this.ej2Instances.templateParser(template);
         },
+        unwireColumnDragEvent(): void {
+            return this.ej2Instances.unwireColumnDragEvent();
+        },
         updateCard(cardData: Object | Object[], index?: number): void {
             return this.ej2Instances.updateCard(cardData, index);
+        },
+        wireColumnDragEvent(): void {
+            return this.ej2Instances.wireColumnDragEvent();
         },
     }
 });
@@ -196,7 +202,9 @@ export type KanbanComponent = typeof ComponentBase & {
     showColumn(key: string | number): void;
     showSpinner(): void;
     templateParser(template: string | Object): any;
-    updateCard(cardData: Object | Object[], index?: number): void
+    unwireColumnDragEvent(): void;
+    updateCard(cardData: Object | Object[], index?: number): void;
+    wireColumnDragEvent(): void
 };
 
 export const KanbanPlugin = {