+++ components/kanban/CHANGELOG.md
@@ -2,13 +2,15 @@
 
 ## [Unreleased]
 
-## 31.2.18 (2025-12-09)
+## 32.1.19 (2025-12-16)
 
 ### Kanban
 
-#### Bug Fixes
+#### New Features
 
-- `#I739684` – Now, Kanban component works properly when the height is set to `100%` with `enableVirtualization` is set to true.
+- **Column Drag and Drop**: This feature allows users to reorder columns in the Kanban control by setting the [`allowColumnDragAndDrop`](https://ej2.syncfusion.com/documentation/api/kanban/#allowColumnDragAndDrop) property to true. When enabled, columns can be dragged and dropped to rearrange their order effortlessly. Explore the demos [here](https://ej2.syncfusion.com/demos/#/material3/kanban/column-rearrangement.html).
+
+- **Column Drag and Drop**: This feature allows users to reorder columns in the Kanban control by setting the [`allowColumnDragAndDrop`](https://ej2.syncfusion.com/vue/documentation/api/kanban/#allowColumnDragAndDrop) property to true. When enabled, columns can be dragged and dropped to rearrange their order effortlessly. Explore the demos [here](https://ej2.syncfusion.com/vue/demos/#/material3/kanban/column-rearrangement.html).
 
 ## 25.1.35 (2024-03-15)